requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.7.0rc1
Django==3.0.7
django-braces==1.14.0
django-countries==6.1.2
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
ecdsa==0.14.1
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-jose==3.2.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
